move CGAL::Mesh_3::internal::Polyline class to its own header
